+++ RAM_Adapter/Delete/_Delete.cs
