+++ RAM_Adapter/RAM_Adapter.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -121,13 +121,29 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Convert\ToBHoM.cs" />
+    <Compile Include="Query\IsColumn.cs" />
+    <Compile Include="Query\GetStory.cs" />
     <Compile Include="Convert\ToRAM.cs" />
-    <Compile Include="CRUD\Create.cs" />
-    <Compile Include="CRUD\Delete.cs" />
+    <Compile Include="Create\Load.cs" />
+    <Compile Include="Create\_Create.cs" />
+    <Compile Include="Create\Bar.cs" />
+    <Compile Include="Create\Material.cs" />
+    <Compile Include="Create\Panel.cs" />
+    <Compile Include="Create\Level.cs" />
+    <Compile Include="Create\Grid.cs" />
+    <Compile Include="Create\Error.cs" />
+    <Compile Include="Delete\_Delete.cs" />
     <Compile Include="AdapterActions\Pull.cs" />
     <Compile Include="AdapterActions\Push.cs" />
-    <Compile Include="CRUD\Read.cs" />
-    <Compile Include="CRUD\UpdateObjects.cs" />
+    <Compile Include="Read\Bar.cs" />
+    <Compile Include="Read\Material.cs" />
+    <Compile Include="Read\Panel.cs" />
+    <Compile Include="Read\Load.cs" />
+    <Compile Include="Read\Grid.cs" />
+    <Compile Include="Read\Level.cs" />
+    <Compile Include="Read\Results.cs" />
+    <Compile Include="Read\_Read.cs" />
+    <Compile Include="Update\UpdateObjects.cs" />
     <Compile Include="RAMAdapter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Types\Comparer.cs" />
