+++ RAM_Adapter/Update/UpdateObjects.cs
